Export BookForm from Book component
